Rename config_loader to data_loader
