+++ test/data_loader_tests.js
@@ -1,0 +1,13 @@
+'use strict';
+
+const tap = require('tap');
+
+tap.test('loadDataFromFile - Load ', function(test){
+  const fs = require('fs-extra'),
+    dataLoader = require('../core/lib/data_loader')(),
+    data_dir = './test/files/_data/';
+
+  let data = dataLoader.loadDataFromFile(data_dir + 'foo', fs);
+  test.equals(data.foo, 'bar');
+  test.end();
+});
